+++ src/CERES/BI/Util/Random.hs
@@ -1,0 +1,7 @@
+module CERES.BI.Util.Random
+  ( module Random
+  )
+where
+
+
+import           Util.Wrapper.Random.SplitMix  as Random